requirements.txt
streamlit>=1.42.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
spacy>=3.8.0
sentence-transformers>=3.0.1
pdfplumber>=0.11.4
PyPDF2>=3.0.1
pandas>=2.2.3
scikit-learn>=1.5.2
openai>=1.54.0
google-genai>=1.12.1
python-docx>=1.1.2
